requirements.txt
aiogram>=3.13.0
python-dotenv>=1.0.0
pymysql>=1.1.0
cryptography>=42.0.0
